requirements.txt
langgraph==0.2.34
qdrant-client==1.9.2
sentence-transformers==3.0.1
pyyaml==6.0.2
requests==2.32.3
openai==1.40.3
flask==2.3.3